apps_capability/requirements.txt
asn1crypto==0.24.0
cffi==1.12.2
cryptography==2.6.1
Django==1.11
django-rest-knox==3.6.0
djangorestframework==3.9.2
enum34==1.1.6
ipaddress==1.0.22
mysqlclient==1.4.2
Pillow==5.4.1
pycparser==2.19
pyOpenSSL==19.0.0
pytz==2018.9
six==1.12.0
